requirements.txt
aws-cdk-lib>=2.179.0
constructs>=10.4.2
boto3>=1.35.83
mypy-boto3-s3>=1.35.93
mypy-boto3-lambda>=1.36.22
cdklabs-generative-ai-cdk-constructs>=0.1.295